Create components: CreateAccount, DeliveryAddress, PaymentDetails
